Update av_club_graph.rb
The "press the number corresponding with the choice you want" approach sounds very customer service over the phone-esque, and if we decide to move forward with Rails, it might be a pain to introduce a new web page to handle that. Right now, I'm messing with different string parsing ideas to determine the choice a user intended from what their input string is. At first, I had the length comparison thing above (commented out) but then realized I could just take the 2 strings, remove all spaces, and compare the characters. Probably a better way of doing it.